Add rudimentary user gating
The user gate is a rudimentary authorization filter that controls which
authenticated users are authorized to make requests against CPS. It is
dynamically configured by a secret containing rules in the kcp-system
namespace of the root workspace.

This filter is meant to be used along with another system that
dynamically updates the secret during a user sign-up workflow ahead of
general availability. It should not be used after a more general
mechanism provided by CIAM is in place.